Bug fixes. Now saves indices. Hopefully improved protons memory management

+++ selection/makeDataSets.py
@@ -1,6 +1,6 @@
 '''
 
-makeDataSets.py  v0.1
+makeDataSets.py  v0.2
 
 Merges together the output of selection.py and creates training/validation/testing sets. Takes a lot of memory!
 
@@ -20,6 +20,7 @@ import ast
 import argparse
 import os
 import time
+import cPickle as pickle
 
 
 def getNrEvents(filelist):
@@ -69,6 +70,12 @@ def getSetIndexes(nrofe,nrofp,trainingFraction,validationFraction,validationMixt
 	
 	i.e. if the list returned is [4,9,15], it means I use events number 4,9, and 15.
 	'''
+	
+	if os.path.isfile('indices.pick'):
+		with open('indices.pick','r') as f:
+			a = pickle.load(f)
+		return a[0], a[1], a[2], a[3], a[4], a[5]
+	
 	# Training
 	available_E = range(nrofe)
 	available_P = range(nrofp)
@@ -106,7 +113,10 @@ def getSetIndexes(nrofe,nrofp,trainingFraction,validationFraction,validationMixt
 		for i in xrange(testMixture * len(selectedE_test)):
 			j = random.randint(0,len(available_P)-1)
 			selectedP_test.append( available_P.pop(j) )
-
+	
+	with open('indices.pick','w') as f:
+		pickle.dump([selectedE_train, selectedP_train, selectedE_validate, selectedP_validate, selectedE_test, selectedP_test],f)
+	
 	return selectedE_train, selectedP_train, selectedE_validate, selectedP_validate, selectedE_test, selectedP_test
 
 
@@ -162,7 +172,7 @@ if __name__ == '__main__':
 	del arr
 	labels = getLabels()
 	
-	print "Got labels (", str(strftime('%H:%M:%S', gmtime( time.time() - t0 ))), ')'
+	print "Got labels (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
 	
 	parser = argparse.ArgumentParser()
 	parser.add_argument("-ts", "--train_split", help="Fraction of total electrons that go into training [0;1]",default=0.6,type=float)
@@ -183,10 +193,10 @@ if __name__ == '__main__':
 	
 	selectedE_train, selectedP_train, selectedE_validate, selectedP_validate, selectedE_test, selectedP_test = getSetIndexes(nrofe,nrofp,trainingFraction,validationFraction,validationMixture,testMixture)
 	
-	print "Got indices (", str(strftime('%H:%M:%S', gmtime( time.time() - t0 ))), ')'
+	print "Got indices (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
 	
 	# Merging and splitting - electrons
-	if not os.path.isfile('dataset_electron_test.root'):
+	if not os.path.isfile('dataset_electrons_test.root'):
 		
 		print "--- Electrons ---"
 		
@@ -197,13 +207,13 @@ if __name__ == '__main__':
 				i=i+1
 				continue
 			arr_e = np.concatenate( (arr_e , np.load(f) ) )
-		print "Built the large array (", str(strftime('%H:%M:%S', gmtime( time.time() - t0 ))), ')'
+		print "Built the large array (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
 		np.random.shuffle(arr_e)
 		set_e_train = arr_e[ selectedE_train, :]
 		set_e_validate = arr_e[ selectedE_validate, :]
 		set_e_test = arr_e[ selectedE_test, :]
 		del arr_e
-		print "Saving train, validate, test arrays (", str(strftime('%H:%M:%S', gmtime( time.time() - t0 ))), ')'
+		print "Saving train, validate, test arrays (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
 		np.save('data_train_elecs.npy',set_e_train)
 		np.save('data_validate_elecs.npy',set_e_validate)
 		np.save('data_test_elecs.npy',set_e_test)
@@ -211,7 +221,7 @@ if __name__ == '__main__':
 		np2root(set_e_validate,getLabels(),outname='dataset_electrons_validate.root')
 		np2root(set_e_test,getLabels(),outname='dataset_electrons_test.root')
 		del set_e_train, set_e_validate, set_e_test
-		print "Done saving (", str(strftime('%H:%M:%S', gmtime( time.time() - t0 ))), ')'
+		print "Done saving (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
 
 	
 	# Protons
@@ -226,21 +236,32 @@ if __name__ == '__main__':
 				i=i+1
 				continue
 			arr_p = np.concatenate( (arr_p , np.load(f) ) )
-		print "Built the large array (", str(strftime('%H:%M:%S', gmtime( time.time() - t0 ))), ')'
+		print "Built the large array (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
 		np.random.shuffle(arr_p)
-		set_p_train = arr_p[ selectedP_train, :]
-		set_p_validate = arr_p[ selectedP_validate, :]
-		set_p_test = arr_p[ selectedP_test, :]
+		
+		"Saving train (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
+		if not os.path.isfile('dataset_protons_train.root'):
+			set_p_train = arr_p[ selectedP_train, :]
+			np.save('data_training_prots.npy',set_p_train)
+			np2root(set_p_train,getLabels(),outname='dataset_protons_train.root')
+			del set_p_train
+		
+		"Saving validate (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
+		if not os.path.isfile('dataset_protons_validate.root'):
+			set_p_validate = arr_p[ selectedP_validate, :]
+			np.save('data_validate_prots.npy',set_p_validate)
+			np2root(set_p_validate,getLabels(),outname='dataset_protons_validate.root')
+			del set_p_validate
+		
+		"Saving test (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
+		it not os.path.isfile('dataset_protons_test.root'):
+			set_p_test = arr_p[ selectedP_test, :]
+			np.save('data_testing_prots.npy',set_p_test)
+			np2root(set_p_test,getLabels(),outname='dataset_protons_test.root')
+			del set_p_test
+		
 		del arr_p
-		print "Saving train, validate, test arrays (", str(strftime('%H:%M:%S', gmtime( time.time() - t0 ))), ')'
-		np.save('data_training_prots.npy',set_p_train)
-		np.save('data_validate_prots.npy',set_p_validate)
-		np.save('data_testing_prots.npy',set_p_test)
-		np2root(set_p_train,getLabels(),outname='dataset_protons_train.root')
-		np2root(set_p_validate,getLabels(),outname='dataset_protons_validate.root')
-		np2root(set_p_test,getLabels(),outname='dataset_protons_test.root')
-		del set_p_train, set_p_validate, set_p_test
-		print "Done saving (", str(strftime('%H:%M:%S', gmtime( time.time() - t0 ))), ')'
+		print "Done saving (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
 	
 	# Concatenate electrons and protons
 	print "Building train set..."
@@ -251,7 +272,7 @@ if __name__ == '__main__':
 	np.save('dataset_train.npy',train_set)
 	np2root(train_set,getLabels(),outname='dataset_train.root')
 	del train_set, set_e_train, set_p_train
-	print "Done  (", str(strftime('%H:%M:%S', gmtime( time.time() - t0 ))), ')'
+	print "Done  (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
 	
 	print "Building validation set..."
 	set_e_validate = np.load('data_validate_elecs.npy')
@@ -261,7 +282,7 @@ if __name__ == '__main__':
 	np.save('dataset_validate.npy',validate_set)
 	np2root(validate_set,getLabels(),outname='dataset_validate.root')
 	del validate_set, set_e_validate, set_p_validate
-	print "Done  (", str(strftime('%H:%M:%S', gmtime( time.time() - t0 ))), ')'
+	print "Done  (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
 	
 	print "Building test set..."
 	set_e_test = np.load('data_test_elecs.npy')
@@ -271,4 +292,4 @@ if __name__ == '__main__':
 	np.save('dataset_test.npy',test_set)
 	np2root(test_set,getLabels(),outname='dataset_test.root')
 	del test_set, set_e_test, set_p_test
-	print "Done (", str(strftime('%H:%M:%S', gmtime( time.time() - t0 ))), ')'
+	print "Done (", str(time.strftime('%H:%M:%S', time.gmtime( time.time() - t0 ))), ')'
